% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contrasts.R
\name{calcDeltaMethodVariance}
\alias{calcDeltaMethodVariance}
\title{Calculates the estimate and variance for causal effects}
\usage{
calcDeltaMethodVariance(vcov1, vcov2, grab_geex_num1, grab_geex_num2,
  contrast_type)
}
\arguments{
\item{vcov1}{the vcov matrix for first population mean}

\item{vcov2}{the vcov matrix for second population mean}

\item{grab_geex_num1}{the correct from of vcov1}

\item{grab_geex_num2}{the correct from of vcov2}

\item{contrast_type}{e.g. "difference" or "negative risk ratio"}
}
\value{
the estimated variance of the causal effect estimate
}
\description{
Calculates the estimate and variance for causal effects
}
