% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{coef.ipw_interference}
\alias{coef.ipw_interference}
\title{Returns the model coefficients from ipw_interference object}
\usage{
\method{coef}{ipw_interference}(object, ...)
}
\arguments{
\item{object}{object of class "ipw_interference"}

\item{...}{dots}
}
\description{
Returns the model coefficients from ipw_interference object
}
