% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contasts.R
\name{contrastThese}
\alias{contrastThese}
\title{Plumber for defining a contrast matrix}
\usage{
contrastThese(x1, x2, contrast_type)
}
\arguments{
\item{x1}{first value}

\item{x2}{second value}

\item{contrast_type}{e.g. "difference" or "negative risk ratio"}
}
\description{
For \code{\link{calcDeltaMethodVariance}}.
}
