% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTV_IPTW.R
\name{eeFunTV_IPTW}
\alias{eeFunTV_IPTW}
\title{Estimating Function for IPTW}
\usage{
eeFunTV_IPTW(data, trt_model_obj, num_fixefs, var_names, alpha,
  x_levels = NULL, integrate_alphas, randomization_probability, weight_type)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{trt_model_obj}{The fitted model object (usually a glm).}

\item{num_fixefs}{Number of fixed effect parameters from treatment model.
Perhaps unncessaary coding.}

\item{var_names}{A list of names for outcome, treatment, clustering, and
perhaps participation.}

\item{alpha}{One of the allocations at a time}

\item{x_levels}{default NULL unless there are factos in design matrix. From
\code{\link[geex]{grab_design_levels}}.}

\item{integrate_alphas}{true}

\item{randomization_probability}{usually 1. e.g. 2/3  in Perez-Heydrich et
al. (2014) Biometrics}

\item{weight_type}{Currently only supports "unstabilized"}
}
\description{
This function is to be passed into geex::m_estimate
}
