% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTargets.R
\name{estimateTargets}
\alias{estimateTargets}
\title{Calcluates point estimates for target estimands}
\usage{
estimateTargets(alphas, num_alphas, num_clusters, unique_clusters,
  grouping_vector, treatment_vector, ps_model_matrix, outcome_vector, fixefs,
  sigma, integrate_alphas, randomization_probability, weight_type)
}
\arguments{
\item{alphas}{the range of allocations or policies from 0 to 1.

  When \code{model_method == "oracle"} then \code{model_options} must be a
  list with named numeric vectors \code{fixefs} and \code{var_comp}. See
  \code{\link{prepareOracle}}. For mixed effect model, note that that random
  intercept's term in the modeling formula (e.g., \code{ ( 1 | cluster_ID )
  }) must be omitted from \code{formula}.

  Arguments that can be passed through \code{...} include: \itemize{ \item
  \code{integrate_alphas}. Not yet supported. \item \code{verbose}. Set to
  \code{TRUE} for more verbose messaging. Default \code{FALSE}. \item
  \code{contrast_type}. Not yet supported. \item \code{keep_components}. Set
  to \code{TRUE} for more verbose output. Default \code{FALSE}. \item
  \code{target_grids}. User can supply target estimands with
  \code{makeTargetGrids}. \item \code{deriv_control}. User can supply the
  \code{deriv_control} argument to \code{\link[geex]{m_estimate}} with
  \code{\link[geex]{setup_deriv_control}}. }}

\item{num_alphas}{The number of allocation parameters}

\item{num_clusters}{The number of unique clusters (i.e., i.i.d. sample units)}

\item{unique_clusters}{The ID values for the unique clusters}

\item{grouping_vector}{The vector of cluster ID for all individuals}

\item{treatment_vector}{The vector of treatment identifiers for all
individuals}

\item{ps_model_matrix}{The matrix of pre-treatment variables in the
propensity score model for all individuals}

\item{outcome_vector}{The vector of observed outcome for all individuals}

\item{fixefs}{The estimated values of the fixed effects parameters from the
propensity score model}

\item{sigma}{The estimated value of the (single) random effect variance
component from the propensity score model}

\item{integrate_alphas}{Optional argument passed from
\code{\link{estimateEffects}}}

\item{randomization_probability}{Optional argument passed from
\code{\link{estimateEffects}}. Usually 1. For example, 2/3  in
Perez-Heydrich et al. (2014) Biometrics.}

\item{weight_type}{Estimators as presented in Liu, Hudgens, and Becker-Dreps
(2016) Biometrika. Select \code{"HT"} for unstabilized weights. Select
\code{"Hajek1"} or \code{"Hajek2"} for stabilized weights. Select
\code{"HT_TV"} for the estimators presented in Tchetgen Tchetgen and
VanderWeele (2012) SMMR and Perez-Heydrich et al. (2014) Biometrics, which
in general target estimands different from those in Liu, Hudgens, and
Becker-Dreps (2016) Biometrika.}
}
\description{
This function calculates the estimated values in a direct plug-in method.
}
