% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTV_IPTW.R
\name{estimateTV_IPTW}
\alias{estimateTV_IPTW}
\title{Stabilized IPTW Estimators for Causal Effects Assuming Partial Interference}
\usage{
estimateTV_IPTW(data, formula, alphas, weight_type = c("HT", "Hajek1",
  "Hajek2", "HT_TV")[3], model_method = c("glm", "glmer", "oracle")[2],
  model_options = list(nAGQ = 5, family = "binomial"), ...)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{formula}{Multi-part formula: \code{Outcome | Treatment ~
model_predictors | cluster_ID}. This will be coerced to object of type
\code{\link[Formula]{Formula}}. When using \code{model_method = "glmer"},
then the random intercept term is supplied in the style preferred by
\code{lme4}'s \code{merMod}: e.g., \code{Outcome | Treatment ~
model_predictors + ( 1 | cluster_ID ) | cluster_ID}.}

\item{alphas}{the range of allocations or policies from 0 to 1.

  When \code{model_method == "oracle"} then \code{model_options} must be a
  list with named numeric vectors \code{fixefs} and \code{var_comp}. See
  \code{\link{prepareOracle}}. For mixed effect model, note that that random
  intercept's term in the modeling formula (e.g., \code{ ( 1 | cluster_ID )
  }) must be omitted from \code{formula}.

  Arguments that can be passed through \code{...} include: \itemize{ \item
  \code{integrate_alphas}. Not yet supported. \item \code{verbose}. Set to
  \code{TRUE} for more verbose messaging. Default \code{FALSE}. \item
  \code{contrast_type}. Not yet supported. \item \code{keep_components}. Set
  to \code{TRUE} for more verbose output. Default \code{FALSE}. \item
  \code{target_grids}. User can supply target estimands with
  \code{makeTargetGrids}. \item \code{deriv_control}. User can supply the
  \code{deriv_control} argument to \code{\link[geex]{m_estimate}} with
  \code{\link[geex]{setup_deriv_control}}. }}

\item{weight_type}{Estimators as presented in Liu, Hudgens, and Becker-Dreps
(2016) Biometrika. Select \code{"HT"} for unstabilized weights. Select
\code{"Hajek1"} or \code{"Hajek2"} for stabilized weights. Select
\code{"HT_TV"} for the estimators presented in Tchetgen Tchetgen and
VanderWeele (2012) SMMR and Perez-Heydrich et al. (2014) Biometrics, which
in general target estimands different from those in Liu, Hudgens, and
Becker-Dreps (2016) Biometrika.}

\item{model_method}{\code{"glm"} for logistic, \code{"glmer"} for logistic
with single random intercept. \code{"oracle"} supported; see
\code{\link[inferference]{interference}} for syntax.}

\item{model_options}{passed to \code{\link[lme4]{glmer}} or perhaps glm(in
future).}

\item{...}{additional args. See details.}
}
\description{
This fits some of the IPW estimators introduced in Liu, Hudgens, and
Becker-Dreps (2016) Biometrika. These estimators estimate causal effects in
the presence of partial interference, with estimates of the asymptotic
variance from standard M-estimation theory.
}
\details{
Note that these estimators estimate different causal estimands than those in
Tchetgen Tchetgen and VanderWeele (2012) SMMR that were applied in
Perez-Heydrich et al. (2014) Biometrics and implemented with
\code{\link[inferference]{interference}} by Saul and Hudgens (2017) JSS.
}
