% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_combined.R
\name{estimateVarianceCombined}
\alias{estimateVarianceCombined}
\title{Estimates of the asymptotic variance of the estimators}
\usage{
estimateVarianceCombined(alphas, num_alphas, mu_alphas_ests, data, num_fixefs,
  fixefs, sigma, x_levels, trt_model_obj, var_names, target_grids,
  randomization_probability, weight_type, verbose, keep_components,
  compute_roots, integrate_alphas, deriv_control, contrast_type)
}
\arguments{
\item{alphas}{the range of allocations or policies from 0 to 1.

  When \code{model_method == "oracle"} then \code{model_options} must be a
  list with named numeric vectors \code{fixefs} and \code{var_comp}. See
  \code{\link{prepareOracle}}. For mixed effect model, note that that random
  intercept's term in the modeling formula (e.g., \code{ ( 1 | cluster_ID )
  }) must be omitted from \code{formula}.

  Arguments that can be passed through \code{...} include: \itemize{ \item
  \code{integrate_alphas}. Not yet supported. \item \code{verbose}. Set to
  \code{TRUE} for more verbose messaging. Default \code{FALSE}. \item
  \code{contrast_type}. Not yet supported. \item \code{keep_components}. Set
  to \code{TRUE} for more verbose output. Default \code{FALSE}. \item
  \code{target_grids}. User can supply target estimands with
  \code{makeTargetGrids}. \item \code{deriv_control}. User can supply the
  \code{deriv_control} argument to \code{\link[geex]{m_estimate}} with
  \code{\link[geex]{setup_deriv_control}}. }}

\item{num_alphas}{The number of allocation parameters}

\item{mu_alphas_ests}{Point estimates of population mean target estimands}

\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{num_fixefs}{The number of fixed effect parameters from treatment model}

\item{fixefs}{The estimated values of the fixed effects parameters from the
propensity score model}

\item{sigma}{The estimated value of the (single) random effect variance
component from the propensity score model}

\item{x_levels}{Default NULL unless there are factos in design matrix. From
\code{\link[geex]{grab_design_levels}}.}

\item{trt_model_obj}{Fitted model object from \code{\link{fitModel}}

 When \code{model_method = "oracle"} see \code{\link{prepareOracle}}.

 @return A list including \itemize{
 \item \code{fixefs}
 \item \code{sigma}
 \item \code{ps_model_matrix}
 \item \code{x_levels}
 }}

\item{var_names}{A list of names for outcome, treatment, clustering, and
perhaps participation.}

\item{target_grids}{output from \code{\link{makeTargetGrids}}}

\item{randomization_probability}{Optional argument passed from
\code{\link{estimateEffects}}. Usually 1. For example, 2/3  in
Perez-Heydrich et al. (2014) Biometrics.}

\item{weight_type}{Estimators as presented in Liu, Hudgens, and Becker-Dreps
(2016) Biometrika. Select \code{"HT"} for unstabilized weights. Select
\code{"Hajek1"} or \code{"Hajek2"} for stabilized weights. Select
\code{"HT_TV"} for the estimators presented in Tchetgen Tchetgen and
VanderWeele (2012) SMMR and Perez-Heydrich et al. (2014) Biometrics, which
in general target estimands different from those in Liu, Hudgens, and
Becker-Dreps (2016) Biometrika.}

\item{verbose}{Optional argument from \code{\link{estimateEffects}}}

\item{keep_components}{Optional argument from \code{\link{estimateEffects}}}

\item{compute_roots}{Optional argument from \code{\link{estimateEffects}}}

\item{integrate_alphas}{Optional argument passed from
\code{\link{estimateEffects}}}

\item{deriv_control}{Optional for \code{\link[geex]{m_estimate}}}

\item{contrast_type}{e.g. "difference" or "negative risk ratio"}
}
\description{
These functions carry out the M-estimation procedures for the "sandwich"
variance estimators
}
