% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Estimate the treatment model parameters.}
\usage{
fitModel(data, modeling_formula, model_options, model_method)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{modeling_formula}{This is just the 'inner' part of the multi-part
  formula for estimating treatment on predictors.

 When \code{model_method = "oracle"} see \code{\link{prepareOracle}}.

 @return \code{trt_model_obj}.

 @export}

\item{model_options}{passed to \code{\link[lme4]{glmer}} or perhaps glm(in
future).}

\item{model_method}{\code{"glm"} for logistic, \code{"glmer"} for logistic
with single random intercept. \code{"oracle"} supported; see
\code{\link[inferference]{interference}} for syntax.}
}
\description{
Output of this function will be handled by \code{\link{getModel}}
}
