% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contrasts.R
\name{getContrastVals}
\alias{getContrastVals}
\title{Returns a contrast matrix's values}
\usage{
getContrastVals(contrast_type)
}
\arguments{
\item{contrast_type}{e.g. "difference" or "negative risk ratio"}
}
\description{
For \code{\link{calcDeltaMethodVariance}}
}
