% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contasts.R
\name{getGeexNum}
\alias{getGeexNum}
\title{Plumber to assist with geex-to-TV conversion}
\usage{
getGeexNum(mu_row, trt_char)
}
\arguments{
\item{mu_row}{one row of the pop_means_grid}

\item{trt_char}{"trt1" or "trt2"}
}
\value{
the row number of the geex output to grab
}
\description{
Plumber to assist with geex-to-TV conversion
}
