% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{getModel}
\alias{getModel}
\title{Grabs necessary information from a treatment model.}
\usage{
getModel(trt_model_obj, data)
}
\arguments{
\item{trt_model_obj}{Fitted model object from \code{\link{fitModel}}

 When \code{model_method = "oracle"} see \code{\link{prepareOracle}}.

 @return A list including \itemize{
 \item \code{fixefs}
 \item \code{sigma}
 \item \code{ps_model_matrix}
 \item \code{x_levels}
 }}

\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}
}
\description{
Grabs necessary information from a treatment model.
}
