% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTargetGrids.R
\name{makeTargetGrids}
\alias{makeTargetGrids}
\title{Makes a dataframe "grid" of estimands to estimate}
\usage{
makeTargetGrids(alphas, treatments = c(0, 1))
}
\arguments{
\item{alphas}{the range of allocations or policies from 0 to 1.}

\item{treatments}{defaults to \code{0:1} as in TV2012}
}
\value{
a list of pop_means and effects grids
}
\description{
This function can be called prior to and then passed into
\code{\link{estimateTV_IPTW}} informally by specifying \code{target_grids =
makeTargetGrids(alphas = c(0.4,0.5))} via the \code{\link{...}} argument.
}
\seealso{
The \code{effect_grid} function in the  inferference package
}
\author{
Bradley Saul and Brian G. Barkley
}
