% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareOracle.R
\name{prepareOracle}
\alias{prepareOracle}
\title{A helper function for using oracle estimates for
\code{\link{estimateTV_IPTW}}.}
\usage{
prepareOracle(data, model_options, modeling_formula)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{model_options}{passed to \code{\link[lme4]{glmer}} or perhaps glm(in
future).}

\item{modeling_formula}{see \code{\link{fitModel}}

This function checks whether the oracle parameters passed through
\code{model_options} to \code{\link{estimateTV_IPTW}} are in appropriate
order. User should supply only the formula that would be
  used to model treatment on predictors. See Details for how to use with
  random effect models and \code{\link[lme4]{glmer}}.

  \code{model_options} must be a list with numeric vectors \code{fixefs} and
  \code{var_comp}. The vector \code{fixefs} must have names. This function
  will provide a check that the names of \code{fixefs} are equal to the names
  that are generated by the call to \code{\link[stats]{model.matrix}}.

  For mixed effect model, note that that random intercept's term in the
  modeling formula (e.g., \code{ ( 1 | cluster_ID ) }) must be omitted from
  \code{formula}.}
}
\value{
A list of elements, perhaps including \itemize{
  \item \code{will_it_work}: Boolean. If \code{FALSE} then there are issues
  that need to be fixed.
  \item \code{fixefs} the named numeric vector of fixed effects parameters.
  \item \code{sigma} Either a numeric singleton, or \code{NULL}.
  \item \code{names_fixefs}: the names of \code{model_options$fixefs}.
  \item \code{names_matrix_modeling}: the colnames of the model matrix.
  \item \code{model_matrix}: the model matrix.
  \item \code{x_levels}: for the \code{xlev} argument in \code{\link[stats]{model.frame}}.
  \item \code{model_options}: the input argument
  \item \code{modeling_formula}: the input argument
  }
}
\description{
A helper function for using oracle estimates for
\code{\link{estimateTV_IPTW}}.
}
