% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.ipw_interference}
\alias{print.ipw_interference}
\title{Prints small amount of output from a ipw_interference object}
\usage{
\method{print}{ipw_interference}(x, ...)
}
\arguments{
\item{x}{object of class "ipw_interference"}

\item{...}{dots}
}
\description{
Prints small amount of output from a ipw_interference object
}
