% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contrasts.R
\name{tidyTargets}
\alias{tidyTargets}
\title{Take geex'd variance output and turn it into a tidy grid of target estimates}
\usage{
tidyTargets(target_grids, pop_mean_alphas_list, contrast_type)
}
\arguments{
\item{target_grids}{output from \code{\link{makeTargetGrids}}}

\item{pop_mean_alphas_list}{looped geex output}

\item{contrast_type}{e.g. "difference" or "negative risk ratio"}
}
\description{
Take geex'd variance output and turn it into a tidy grid of target estimates
}
