% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contasts_combined.R
\name{tidyTargetsCombined}
\alias{tidyTargetsCombined}
\title{Take geex'd variance output and turn it into a tidy grid of target estimates}
\usage{
tidyTargetsCombined(target_grids, alpha_grid, geex_output, contrast_type)
}
\arguments{
\item{target_grids}{output from \code{\link{makeTargetGrids}}}

\item{alpha_grid}{estimates and variance estimates for the pop means}

\item{geex_output}{from \code{\link[geex]{m_estimate}}.}

\item{contrast_type}{e.g. "difference" or "negative risk ratio"}

\item{pop_mean_alphas_list}{looped geex output}
}
\description{
Take geex'd variance output and turn it into a tidy grid of target estimates
}
