% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_contrasts_combined.R
\name{trtIdx}
\alias{trtIdx}
\title{Helper function for \code{\link{tidyTargetsCombined}}}
\usage{
trtIdx(trt)
}
\arguments{
\item{trt}{either 1, 0, or NA. The value of \eqn{z} in \eqn{\mu(z,\alpha)}.}
}
\description{
Helper function for \code{\link{tidyTargetsCombined}}
}
