% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTV_IPTW.R
\name{estimateTV_IPTW}
\alias{estimateTV_IPTW}
\title{Estimate ATE with IPTW
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.}
\usage{
estimateTV_IPTW(data, formula, alphas, weight_type = c("HT", "Hajek1",
  "Hajek2")[3], model_method = c("glm", "glmer")[2],
  model_options = list(nAGQ = 5, family = "binomial"), ...)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{formula}{Three-part formula: Outcome | Treatment ~ model_predictors | cluster_ID. Will be coerced to object of type Formula.}

\item{alphas}{the range of allocations or policies from 0 to 1.}

\item{weight_type}{Currently only supports "unstabilized"}

\item{model_method}{currently only supported "logistic" for logit-link binomial GLM.}

\item{model_options}{passed to \code{\link[lme4]{glmer}} or perhaps glm(in future).}

\item{...}{additional args}
}
\description{
Estimate ATE with IPTW
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
