% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_combined.R
\name{estimateVarianceCombined}
\alias{estimateVarianceCombined}
\title{Estimates the variance}
\usage{
estimateVarianceCombined(alphas, num_alphas, mu_alphas_ests, data, num_fixefs,
  fixefs, sigma, x_levels, trt_model_obj, var_names, target_grids,
  randomization_probability, weight_type, verbose, keep_components,
  compute_roots, integrate_alphas, deriv_control, contrast_type)
}
\arguments{
\item{alphas}{the range of allocations or policies from 0 to 1.

  When \code{model_method == "oracle"} then \code{model_options} must be a
  list with named numeric vectors \code{fixefs} and \code{var_comp}. See
  \code{\link{prepareOracle}}. For mixed effect model, note that that random
  intercept's term in the modeling formula (e.g., \code{ ( 1 | cluster_ID )
  }) must be omitted from \code{formula}.

  Arguments that can be passed through \code{...} include: \itemize{ \item
  \code{integrate_alphas}. Not yet supported. \item \code{verbose}. Set to
  \code{TRUE} for more verbose messaging. Default \code{FALSE}. \item
  \code{contrast_type}. Not yet supported. \item \code{keep_components}. Set
  to \code{TRUE} for more verbose output. Default \code{FALSE}. \item
  \code{target_grids}. User can supply target estimands with
  \code{makeTargetGrids}. \item \code{deriv_control}. User can supply the
  \code{deriv_control} argument to \code{\link[geex]{m_estimate}} with
  \code{\link[geex]{setup_deriv_control}}. }}

\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{num_fixefs}{etc

asd



etc}

\item{trt_model_obj}{Fitted model object from \code{\link{fitModel}}

 When \code{model_method = "oracle"} see \code{\link{prepareOracle}}.

 @return A list including \itemize{
 \item \code{fixefs}
 \item \code{sigma}
 \item \code{ps_model_matrix}
 \item \code{x_levels}
 }}

\item{weight_type}{Estimators as presented in Liu, Hudgens, and Becker-Dreps
(2016) Biometrika. Select \code{"HT"} for unstabilized weights. Select
\code{"Hajek1"} or \code{"Hajek2"} for stabilized weights. Select
\code{"HT_TV"} for the estimators presented in Tchetgen Tchetgen and
VanderWeele (2012) SMMR and Perez-Heydrich et al. (2014) Biometrics, which
in general target estimands different from those in Liu, Hudgens, and
Becker-Dreps (2016) Biometrika.}
}
\description{
somethinhg
}
